requirements.txt
scikit-image
PyMuPDF
python-docx
opencv-python
scipy
tensorflow
gdown
pydot
matplotlib
pandas
tqdm
Flask
numpy
gunicorn
requests
Pillow